+++ Assets/Scripts/AI/AI_Soldier_States.cs
@@ -40,7 +40,7 @@ public class AI_Soldier_States : MonoBehaviour
         {
             case State.Idle:
                 // to be expanded later, maybe add patrol functionality
-                if (target != null)
+                /* if (target != null)
                     state = State.ChasingFoe;
                 else if(macroTarget != null)
                 {
@@ -49,15 +49,14 @@ public class AI_Soldier_States : MonoBehaviour
                 }
                 else
                 {
-                    unit.Idling();
                     state = State.SearchingFoe;
-                }
+                }*/
                 break;
 
             case State.SearchingFoe:
                 target = vision.SearchForFoes();
                 if (target == null)
-                    state = State.Returning;
+                    state = State.Idle;
                 else
                     state = State.ChasingFoe;
                 break;